feat: add shabbat api route and refactor parasha fetch
